+++ CHANGELOG.md
@@ -8,6 +8,16 @@
 - Changes
 ```
 
+## [2.21.0] - 2016/12/08
+### Added
+- User access to user content card on user page
+- User guilds to user page
+
+### Changed
+- User account name to be more appropriate when unavailable
+- Sign up button margins
+- Buttons border width to 2px
+
 ## [2.20.1] - 2016/12/02
 ### Fixed
 - Gw2 map component from linking to wiki BETA map that doesn't exist